ai-test/app/pyproject.toml
[tool.poetry]
name = "ai-test"
version = "1.0"
description = "'An application for demonstrating your skills in AI'"
authors = ["Rockstar"]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.11"
openfabric-pysdk = "^0.3.0"
sentence-transformers = "^2.2.2"
torch = "^2.1.0"
streamlit = "^1.29.0"
pillow = "^10.1.0"
numpy = "^1.24.0"
sqlalchemy = "^2.0.0"
faiss-cpu = "^1.7.4"
chromadb = "^0.4.22"
python-dotenv = "^1.0.0"
scikit-learn = "^1.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[[tool.poetry.source]]
name = "node2"
url = "https://repo.node2.openfabric.network/index"
priority = "supplemental"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
